Add card flipping and picking up
For now, picking up a card just moves it closer to the camera (so the
image is on top). Setting down places it above the highest card below
it.

In the future, picking up a card will potentially lift other cards
stacked on top. Then they will be moved as a group.